+++ CHANGES.md
@@ -45,6 +45,7 @@ Changes in CUPS v2.5b1 (TBA)
 - Updated `cupsRasterReadPixels` and `cupsRasterWritePixels` to not try reading
   or writing if the number of bytes passed is 0 (Issue #914)
 - Updated and documented the MIME typing buffering limit (Issue #925)
+- Updated the maximum file descriptor limit for `cupsd` to 64k-1 (Issue #989)
 - Fixed use-after-free in `cupsdAcceptClient()` when we log warning during error
   handling (fixes CVE-2023-34241)
 - Fixed hanging of `lpstat` on Solaris (Issue #156)